Refactor account settings
