+++ StudentUsos/Features/Settings/Views/SettingsPage.xaml.cs
@@ -1,5 +1,4 @@
 ﻿using StudentUsos.Controls;
-using StudentUsos.Features.Authorization.Services;
 using StudentUsos.Resources.LocalizedStrings;
 
 namespace StudentUsos.Features.Settings.Views;
@@ -22,21 +21,7 @@ public partial class SettingsPage : CustomContentPageNotAnimated
         versionLabel.Text = LocalizedStrings.Version + " " + AppInfo.Current.VersionString;
     }
 
-    private void RevokeAccessButton_Clicked(object sender, EventArgs e)
-    {
-        MessagePopupPage.CreateAndShow(LocalizedStrings.SettingsPage_RevokeAccessTitle, LocalizedStrings.SettingsPage_RevokeAccessDescription,
-            LocalizedStrings.Continue, LocalizedStrings.Cancel, ContinueRevokingAccessAsync, () => { });
 
-        async void ContinueRevokingAccessAsync()
-        {
-            await Browser.OpenAsync("https://usosapps.put.poznan.pl/", BrowserLaunchMode.SystemPreferred);
-        }
-    }
-
-    private void LogOutButton_Clicked_1(object sender, EventArgs e)
-    {
-        MessagePopupPage.CreateAndShow(LocalizedStrings.LoggingOut, LocalizedStrings.LoggingOut_Confirmation, LocalizedStrings.Yes, LocalizedStrings.No, () => AuthorizationService.LogoutAsync(), null);
-    }
 
     private void OpenLogsButton_Clicked(object sender, EventArgs e)
     {
