+++ StudentUsos/Features/Settings/Views/Subpages/AccountSubpage.xaml
@@ -1,0 +1,27 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<controls:CustomContentPageNotAnimated
+    x:Class="StudentUsos.Features.Settings.Views.Subpages.AccountSubpage"
+    xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+    xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+    xmlns:controls="clr-namespace:StudentUsos.Controls"
+    xmlns:strings="clr-namespace:StudentUsos.Resources.LocalizedStrings"
+    xmlns:toolkit="http://schemas.microsoft.com/dotnet/2022/maui/toolkit">
+
+    <ScrollView>
+        <StackLayout Margin="30">
+
+            <!--<Label FontSize="14" HorizontalTextAlignment="Center" Text="{x:Static strings:LocalizedStrings.Account}" />-->
+            <Border Margin="0,5,0,20" Padding="0">
+                <StackLayout>
+                    <Button Text="{x:Static strings:LocalizedStrings.MorePage_RevokeAccess}" Padding="15"
+                                Clicked="RevokeAccessButton_Clicked" CornerRadius="0"/>
+                    <StackLayout BackgroundColor="#444" HeightRequest="1" Margin="20,0"></StackLayout>
+                    <Button Text="{x:Static strings:LocalizedStrings.Logout}" Padding="15"
+                                Clicked="LogOutButton_Clicked_1" CornerRadius="0"/>
+                </StackLayout>
+            </Border>
+
+        </StackLayout>
+    </ScrollView>
+
+</controls:CustomContentPageNotAnimated>
